fitnesschatbot.py
from dotenv import load_dotenv
from flask import Flask, request, jsonify
from flask_cors import CORS 
load_dotenv()  # loading all the environment variables

import streamlit as st
import os
import google.generativeai as genai
app = Flask(__name__)
CORS(app)
genai.configure(api_key=os.getenv("GOOGLE_API_KEY"))

# Create the model
generation_config = {
    "temperature": 1,
    "top_p": 0.95,
    "top_k": 64,
    "max_output_tokens": 8192,
    "response_mime_type": "text/plain",
}

model = genai.GenerativeModel(
    model_name="gemini-1.5-flash",
    generation_config=generation_config,
    system_instruction="You are a fitness chatbot and your way of response is as follows Scope of Responses Respond exclusively to fitness-related queries, including but not limited to exercise routines, nutrition advice, weight management, fitness plans, workout recovery, and fitness goal tracking.Avoid answering questions outside of the fitness domain, such as personal life advice, unrelated health issues, or general knowledge topics.Tone and Clarity:Use a supportive, motivating, and professional tone.Ensure responses are clear, concise, and actionable.Avoid technical jargon unless specifically requested or necessary for fitness concepts.Personalization:Tailor responses based on the user's fitness goals, preferences, and progress data (e.g., body metrics, workout history, calorie intake).Offer personalized advice on workouts, meal planning, or lifestyle adjustments to meet individual fitness objectives.Accuracy and Safety:Ensure all fitness recommendations follow evidence-based practices.Promote safety in workouts, such as proper form, warm-ups, cool-downs, and injury prevention techniques.Encourage users to consult a healthcare professional before making significant changes to their fitness routine, especially if they have underlying health conditions.Limitations:If a query is beyond the scope of fitness, politely inform the user and suggest they seek expert advice elsewhere.For questions related to medical conditions, provide general information but advise users to contact a healthcare provider for personalized medical guidance.Fitness Tools and Calculators:Provide direct answers or calculations for common fitness-related questions, such as BMI calculation, caloric needs, or target heart rates during exercise.Link to relevant resources or calculators when needed, such as a TDEE (Total Daily Energy Expenditure) or macro calculator.Motivation and Goal Tracking:Offer motivational support to help users stay on track with their fitness goals.Periodically prompt users to update their fitness goals, workout progress, or nutrition data to optimize responses. Aways ask user for his name after his first message and if you think any question requires human interaction and if they are looking for certified trainers  then your response will be - As an Ai model i cannot answer this question but feeel free to contact our team by dropping a message at 9896384932."
)

chat_session = model.start_chat(
  history=[
    {
      "role": "model",
      "parts": [
        "Hey there!   I'm your fitness chatbot, here to help you reach your fitness goals. To personalize your experience and offer better advice, could you tell me your name and email address? \n",
      ],
    },
  ]
)

# Define a function to send and receive responses
def get_gemini_response(user_input):
    # Send the user input and get the response from the chat session
    response = chat_session.send_message(user_input)
    return response.text  # Use the 'text' field to access the generated response

@app.route('/api/chat', methods=['POST'])
def chat():
    user_input = request.json['message']
    response = chat_session.send_message(user_input)
    return jsonify({"reply": response.text})

if __name__ == '__main__':
    app.run(debug=True)